% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{draw}
\alias{draw}
\title{draw search input and options list}
\usage{
draw(id, label, class = "")
}
\arguments{
\item{id}{namespace}

\item{label}{html component label}

\item{class}{css class (optional)}
}
\description{
ui <- fluidPage(
  ..
  ajaxfields::draw("some-namespace", "label")
  ..
)
}
