% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{observer}
\alias{observer}
\title{shiny server (module) function}
\usage{
observer(input, output, session)
}
\value{
reactive data
}
\description{
shiny server (module) function
}
\examples{
 server <- function(input, output, session) {
   ..
   data <- callModule(ajaxfields::observer, "some-namespace")
   ..
 }
}
