% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{draw}
\alias{draw}
\title{draw search input and options list}
\usage{
draw(id, host, label)
}
\arguments{
\item{id}{namespace}

\item{host}{post db host}

\item{label}{html component label}
}
\description{
ui <- fluidPage(
  ..
  ajaxfields::draw("some-namespace","http://ncbi.post.host/url/")
  ..
)
}
