% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{draw}
\alias{draw}
\title{draw search input and options list}
\usage{
draw(id, host)
}
\arguments{
\item{id}{namespace}

\item{host}{post db host}
}
\description{
draw search input and options list
}
\examples{
 ui <- fluidPage(
   ..
   ajaxfields::draw("some-namespace","http://ncbi.post.host/url/")
   ..
 )
}
