% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{loadEngine}
\alias{loadEngine}
\title{shiny server load modules engine}
\usage{
loadEngine(session, id, engine, url, limit = 1000)
}
\arguments{
\item{session}{shiny server session}

\item{id}{namespace}

\item{engine}{engine type (simple/es)}

\item{url}{engine url}

\item{limit}{results max size (optional, default 1000)}
}
\description{
server <- function(input, output, session) {
  ..
  ajaxfields::loadEngine(session, "some-namespace", "simle", ""http://ncbi.post.host/url/")
  ajaxfields::loadEngine(session, "some-other-namespace", "es", "http://elastic.seach:9200/indexName/")
  ..
}
}
