% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{observer}
\alias{observer}
\title{shiny server (module) function}
\usage{
observer(input, output, session)
}
\arguments{
\item{input}{unused}

\item{output}{unused}

\item{session}{unused}
}
\value{
reactive list
}
\description{
server <- function(input, output, session) {
  ..
  data <- callModule(ajaxfields::observer, "some-namespace")
  ..
}
}
